+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/mainnet/ProtocolSpec.java
@@ -21,6 +21,7 @@ import org.hyperledger.besu.ethereum.core.BlockImporter;
 import org.hyperledger.besu.ethereum.core.Wei;
 import org.hyperledger.besu.ethereum.core.fees.EIP1559;
 import org.hyperledger.besu.ethereum.core.fees.TransactionPriceCalculator;
+import org.hyperledger.besu.ethereum.mainnet.feemarket.FeeMarket;
 import org.hyperledger.besu.ethereum.privacy.PrivateTransactionProcessor;
 import org.hyperledger.besu.ethereum.vm.EVM;
 import org.hyperledger.besu.ethereum.vm.GasCalculator;
@@ -69,6 +70,8 @@ public class ProtocolSpec {
 
   private final TransactionPriceCalculator transactionPriceCalculator;
 
+  private final Optional<FeeMarket> feeMarket;
+
   private final Optional<EIP1559> eip1559;
 
   private final BadBlockManager badBlockManager;
@@ -98,6 +101,7 @@ public class ProtocolSpec {
    * @param skipZeroBlockRewards should rewards be skipped if it is zero
    * @param gasCalculator the gas calculator to use.
    * @param transactionPriceCalculator the transaction price calculator to use.
+   * @param feeMarket an {@link Optional} wrapping {@link FeeMarket} class if appropriate.
    * @param eip1559 an {@link Optional} wrapping {@link EIP1559} manager class if appropriate.
    * @param badBlockManager the cache to use to keep invalid blocks
    * @param powHasher the proof-of-work hasher
@@ -123,6 +127,7 @@ public class ProtocolSpec {
       final boolean skipZeroBlockRewards,
       final GasCalculator gasCalculator,
       final TransactionPriceCalculator transactionPriceCalculator,
+      final Optional<FeeMarket> feeMarket,
       final Optional<EIP1559> eip1559,
       final BadBlockManager badBlockManager,
       final Optional<PoWHasher> powHasher) {
@@ -146,6 +151,7 @@ public class ProtocolSpec {
     this.skipZeroBlockRewards = skipZeroBlockRewards;
     this.gasCalculator = gasCalculator;
     this.transactionPriceCalculator = transactionPriceCalculator;
+    this.feeMarket = feeMarket;
     this.eip1559 = eip1559;
     this.badBlockManager = badBlockManager;
     this.powHasher = powHasher;
@@ -327,8 +333,17 @@ public class ProtocolSpec {
     return eip1559;
   }
 
-  public boolean isEip1559() {
-    return eip1559.isPresent();
+  /**
+   * Returns the Fee Market used in this specification.
+   *
+   * @return the {@link Optional} wrapping Fee Market implementation.
+   */
+  public Optional<FeeMarket> getFeeMarket() {
+    return feeMarket;
+  }
+
+  public boolean hasFeeMarket() {
+    return feeMarket.isPresent();
   }
 
   /**